split AdminController class
Move code related to the product edit page to the separate controller.
